E-013: cast the numeric arguments to current_clusters
The function is declared current_clusters(NUMERIC, INT, NUMERIC, INT).
DEPTH_THRESHOLD_CM and CLUSTER_EPS_FT are Python floats, so psycopg sent double
precision. smallint to integer is an implicit cast and resolved, but double
precision to numeric is an assignment cast and function resolution does not
apply those, so PostgreSQL reported the function as nonexistent. It existed; the
call did not resolve.

Cast at the call site rather than widening the signature. NUMERIC is right for a
depth in centimetres and a distance in feet.

This never failed in tests because the suite is moto throughout and never runs
SQL against PostGIS. First execution against a real database was on the graded
instance, which is worth stating in the report rather than papering over.

+++ curbline/db.py
@@ -138,13 +138,17 @@ def upsert_alert(
 
 def current_clusters() -> list[dict[str, Any]]:
     """Run DBSCAN over currently-inundated sensors and return zone candidates."""
+    # The casts are load-bearing. current_clusters declares NUMERIC for the
+    # depth and distance parameters, psycopg sends Python floats as double
+    # precision, and double precision to numeric is an assignment cast rather
+    # than an implicit one, so PostgreSQL will not resolve the call. See E-013.
     with cursor() as cur:
         cur.execute(
             """
             SELECT cluster_id, sensor_ids, sensor_count, max_depth_cm,
                    ST_AsGeoJSON(hull) AS hull_geojson,
                    alert_for_hull(hull) AS alert_id
-            FROM current_clusters(%s, %s, %s, %s)
+            FROM current_clusters(%s::numeric, %s::int, %s::numeric, %s::int)
             """,
             (config.DEPTH_THRESHOLD_CM, config.READING_WINDOW_MINS,
              config.CLUSTER_EPS_FT, config.CLUSTER_MIN_SENSORS),
